backend/requirements.txt
# Update requirements to use python-jose instead of PyJWT for FastAPI compatibility
# BUJA Core Platform - Backend Dependencies
# Python 3.10+

# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0
python-multipart>=0.0.9

# Database & Storage
redis[hiredis]>=5.0.3
neo4j>=5.18.0
pinecone>=6.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.29.0

# LLM & AI
langchain>=0.1.13
langchain-community>=0.0.29
langchain-openai>=0.1.0
langgraph>=0.0.30
openai>=1.14.0

# HTTP Client
httpx>=0.25.2,<0.28.0
aiohttp>=3.9.1
pypdf>=4.2.0

# Task Queue (Optional - for Gardener)
celery>=5.3.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2

# Telegram Integration
python-telegram-bot==20.7

# Monitoring & Logging
structlog==24.1.0
sentry-sdk==1.40.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
isort==5.13.2
mypy==1.8.0

# Document Rendering
weasyprint>=62.0
markdown>=3.6

# Reference import utilities
openpyxl>=3.1.0
pytesseract>=0.3.13
pdf2image>=1.17.0
